requirements.txt
PyMuPDF>=1.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.26.0
rank-bm25>=0.2.2
ollama>=0.3.0
fastapi>=0.111.0
slowapi>=0.1.9
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-docx>=1.1.0
pytest>=8.0.0
